requirements.txt
Django>=5
psycopg2-binary>=2.9
python-dotenv>=1.0
